Update OneDrive cleaning script
Unconditional removal is dangerous. OneDrive can stop working for various reasons, and after a long time a lot of files from your desktop or documents can accumulate in this folder. Without the /S switch, this command will only delete the folder when it is empty. We don't want users to delete their files!